+++ tests/wpunit/QuantityNumberFieldTest.php
@@ -171,7 +171,10 @@ class QuantityNumberFieldTest extends FormFieldTestCase implements FormFieldTest
 					shouldExport
 				}
 				placeholder
-				productField
+				connectedProductField {
+					databaseId
+					type
+				}
 				type
 				value
 				... on QuantityNumberField {